Removed Seed data. Cleaned up testing on routes

+++ controllers/userController.js
@@ -46,7 +46,10 @@ module.exports = {
         return res.status(404).json({ message: 'No user with that ID' });
       }
 
-      await Thought.deleteMany({ _id: { $in: user.thoughts } });
+      if (user.thoughts.length > 0) {
+        await Thought.deleteMany({ _id: { $in: user.thoughts } });
+      }
+
       res.json({ message: 'User and associated thoughts deleted!' })
     } catch (err) {
       res.status(500).json(err);
